build: add Tepl dependency, second try
First step to be able to use the Tepl library:
https://wiki.gnome.org/Projects/Tepl

as part of this project:
https://wiki.gnome.org/Apps/Gedit/ReusableCode

tepl_req and tepl_dep follow gtksourceview_req/dep: wherever the
gtksourceview one is present, add a line for tepl too.

This was reverted in commit 38f10366a278266645aff2aee45f08ea82146dbb
because gnome-build-meta was not updated beforehand. I thought that it
was fine to update gnome-build-meta afterwards, but nowadays
gnome-build-meta should be able to build everything at all time (which
is perfectly understandable).

https://gitlab.gnome.org/GNOME/gedit/issues/229